Solution of the Dice Combinations problem, Line Segment Intersection problem, Polygon Area problem, Point Location Test problem, and changes in code format of the problems in the Introductory Problems folder

+++ Introductory Problems/TwoKnights.cpp
@@ -1,8 +1,10 @@
 #include <iostream>
 using namespace std;
 #define ll long long
+
 int main() {
-    int n;
+    ll n;
     cin >> n;
-    for (ll k=1; k<=n; k++) cout << (k*k)*(k*k-1)/2 - 4*(k-1)*(k-2) << endl;
+    for (int k = 1; k <= n; k++) 
+        cout << (k * k) * (k * k - 1) / 2 - 4 * (k - 1) * (k - 2) << endl;
 }
